Refactor after the PATCH Entity implementation
