+++ src/lib/orionld/db/dbModelFromApiAttribute.cpp
@@ -1,0 +1,183 @@
+/*
+*
+* Copyright 2022 FIWARE Foundation e.V.
+*
+* This file is part of Orion-LD Context Broker.
+*
+* Orion-LD Context Broker is free software: you can redistribute it and/or
+* modify it under the terms of the GNU Affero General Public License as
+* published by the Free Software Foundation, either version 3 of the
+* License, or (at your option) any later version.
+*
+* Orion-LD Context Broker is distributed in the hope that it will be useful,
+* but WITHOUT ANY WARRANTY; without even the implied warranty of
+* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero
+* General Public License for more details.
+*
+* You should have received a copy of the GNU Affero General Public License
+* along with Orion-LD Context Broker. If not, see http://www.gnu.org/licenses/.
+*
+* For those usages not covered by this license please contact with
+* orionld at fiware dot org
+*
+* Author: Ken Zangelin
+*/
+extern "C"
+{
+#include "kbase/kMacros.h"                                       // K_VEC_SIZE
+#include "kalloc/kaStrdup.h"                                     // kaStrdup
+#include "kjson/KjNode.h"                                        // KjNode
+#include "kjson/kjLookup.h"                                      // kjLookup
+#include "kjson/kjBuilder.h"                                     // kjString, kjObject, kjChildAdd, ...
+}
+
+#include "logMsg/logMsg.h"                                       // LM_*
+
+#include "orionld/common/orionldState.h"                         // orionldState
+#include "orionld/common/orionldError.h"                         // orionldError
+#include "orionld/common/dotForEq.h"                             // dotForEq
+#include "orionld/kjTree/kjArrayAdd.h"                           // kjArrayAdd
+#include "orionld/kjTree/kjTimestampAdd.h"                       // kjTimestampAdd
+#include "orionld/kjTree/kjTreeLog.h"                            // kjTreeLog
+#include "orionld/db/dbModelFromApiSubAttribute.h"               // dbModelFromApiSubAttribute
+#include "orionld/db/dbModelFromApiAttribute.h"                  // Own interface
+
+
+
+// -----------------------------------------------------------------------------
+//
+// dbModelFromApiAttribute -
+//
+// FOREACH attribute:
+//   * First an object "md" is created, and all fields of the attribute, except the special ones are moved inside "md".
+//     Special fields:
+//     - type
+//     - value/object/languageMap (must be renamed to "value" - that's part of the database model)
+//     - observedAt
+//     - datasetId
+//     - unitCode
+//
+//   * Actually, all children of 'attrP' are moved to 'md'
+//   * And then, all the special fields are moved back to 'attrP'
+//   * ".added"   is added
+//   * ".removed" is added
+//   * "modDate"  is added
+//   * "creDate"  is added iff the attribute did not previously exist
+//
+bool dbModelFromApiAttribute(KjNode* attrP, KjNode* dbAttrsP, KjNode* attrAddedV, KjNode* attrRemovedV)
+{
+  KjNode* mdP         = NULL;
+  char*   attrDotName = kaStrdup(&orionldState.kalloc, attrP->name);  // Needed for attrAddedV, attrRemovedV
+
+  dotForEq(attrP->name);  // Orion DB-Model states that dots are replaced for equal signs in attribute names
+
+  if (attrP->type == KjNull)
+  {
+    LM_TMP(("KZ: '%s' has a null RHS - to be removed", attrP->name));
+    KjNode* attrNameP = kjString(orionldState.kjsonP, NULL, attrDotName);
+    kjChildAdd(attrRemovedV, attrNameP);
+    return true;
+  }
+
+  // Move everything into "md", leaving attrP EMPTY
+  mdP = kjObject(orionldState.kjsonP, "md");
+  mdP->value.firstChildP   = attrP->value.firstChildP;
+  mdP->lastChild           = attrP->lastChild;
+  attrP->value.firstChildP = NULL;
+  attrP->lastChild         = NULL;
+
+  // Move special fields back to "attrP"
+  const char* specialV[] = { "type", "value", "object", "languageMap", "datasetId" };  // observedAt+unitCode are mds (db-model)
+  for (unsigned int ix = 0; ix < K_VEC_SIZE(specialV); ix++)
+  {
+    KjNode* nodeP = kjLookup(mdP, specialV[ix]);
+    if (nodeP != NULL)
+    {
+      if ((ix == 2) || (ix == 3))         // "object", "languageMap", change name to "value" (Orion's DB model)
+        nodeP->name = (char*) "value";
+
+      kjChildRemove(mdP, nodeP);
+      kjChildAdd(attrP, nodeP);
+    }
+  }
+
+  // Also, the attribute is being modified - need to update the "modifiedAt" (called modDate in Orion's DB-Model)
+  kjTimestampAdd(attrP, "modDate");
+
+  KjNode*  mdAddedP     = NULL;
+  KjNode*  mdRemovedP   = NULL;
+  KjNode*  mdNamesP     = NULL;
+  KjNode*  dbAttrP      = kjLookup(dbAttrsP, attrP->name);
+  KjNode*  dbMdP        = NULL;
+
+  if (dbAttrP == NULL)  // Attribute does not already exist
+  {
+    if (attrP->type == KjNull)
+    {
+      LM_W(("Attempt to DELETE an attribute that doesn't exist (%s)", attrP->name));
+      orionldError(OrionldResourceNotFound, "Cannot delete an attribute that does not exist", attrP->name, 404);
+      // Add to 207
+      return false;
+    }
+
+    kjTimestampAdd(attrP, "creDate");
+
+    // The "attrNames" array stores the sub-attribute names in their original names, with dots - attrDotName
+    kjChildAdd(attrAddedV, kjString(orionldState.kjsonP, NULL, attrDotName));
+
+    // All attributes have a field 'mdNames' - it's an array
+    mdNamesP = kjArray(orionldState.kjsonP, "mdNames");
+    kjChildAdd(attrP, mdNamesP);
+
+    LM_TMP(("KX: the attribute is NEW - if it has any sub-attributes, the entire 'md' needs to be included, never mind .added, .removed etc"));
+  }
+  else
+  {
+    // The attribute is not new, but, perhaps it didn't have any sub-attrs before but now it does?
+    // If so, never mind .added, .removed etc, what we need is to include the entire "md"
+    //
+    dbMdP    = kjLookup(dbAttrP, "md");
+    mdNamesP = kjLookup(dbAttrP, "mdNames");
+
+    if (mdNamesP == NULL)
+    {
+      //
+      // No 'mdNames' field in DB Attribute?
+      // That's a DB Error.
+      // Never mind, let's just silently fix it
+      //
+      mdNamesP = kjArray(orionldState.kjsonP, ".names");
+    }
+    else
+    {
+      kjChildRemove(dbAttrP, mdNamesP);
+      mdNamesP->name = (char*) ".names";  // We hijack the 'mdNames' array for the compilation of the Patch Tree
+    }
+
+    kjChildAdd(attrP, mdNamesP);
+
+    mdAddedP   = kjArrayAdd(attrP, ".added");    // FIXME: Should not be needed if dbMdP == NULL
+    mdRemovedP = kjArrayAdd(attrP, ".removed");  // FIXME: Should not be needed if dbMdP == NULL
+  }
+
+  kjTreeLog(attrP, "KZ4: Attribute with unprocessed sub-attributes");
+  if (attrP->type == KjNull)
+    return true;
+
+  // Sub-Attributes
+  LM_TMP(("KZ: Loop over sub-attributes (contents of 'mdP')"));
+  for (KjNode* subP = mdP->value.firstChildP; subP != NULL; subP = subP->next)
+  {
+    LM_TMP(("KZ5: sub-attribute '%s'", subP->name));
+    dbModelFromApiSubAttribute(subP, dbMdP, mdAddedP, mdRemovedP);
+  }
+
+  if (mdP->value.firstChildP != NULL)
+  {
+    LM_TMP(("KY: putting 'md' inside attr"));
+    kjChildAdd(attrP, mdP);
+  }
+  kjTreeLog(attrP, "KY: Attribute with PROCESSED sub-attributes");
+
+  return true;
+}
