+++ src/lib/orionld/db/dbModelFromApiSubAttribute.cpp
@@ -1,0 +1,147 @@
+/*
+*
+* Copyright 2022 FIWARE Foundation e.V.
+*
+* This file is part of Orion-LD Context Broker.
+*
+* Orion-LD Context Broker is free software: you can redistribute it and/or
+* modify it under the terms of the GNU Affero General Public License as
+* published by the Free Software Foundation, either version 3 of the
+* License, or (at your option) any later version.
+*
+* Orion-LD Context Broker is distributed in the hope that it will be useful,
+* but WITHOUT ANY WARRANTY; without even the implied warranty of
+* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero
+* General Public License for more details.
+*
+* You should have received a copy of the GNU Affero General Public License
+* along with Orion-LD Context Broker. If not, see http://www.gnu.org/licenses/.
+*
+* For those usages not covered by this license please contact with
+* orionld at fiware dot org
+*
+* Author: Ken Zangelin
+*/
+extern "C"
+{
+#include "kalloc/kaStrdup.h"                                     // kaStrdup
+#include "kjson/KjNode.h"                                        // KjNode
+#include "kjson/kjLookup.h"                                      // kjLookup
+#include "kjson/kjBuilder.h"                                     // kjString, kjChildAdd, ...
+}
+#include "logMsg/logMsg.h"                                       // LM_*
+
+#include "orionld/common/orionldState.h"                         // orionldState
+#include "orionld/common/orionldError.h"                         // orionldError
+#include "orionld/common/dotForEq.h"                             // dotForEq
+#include "orionld/kjTree/kjTimestampAdd.h"                       // kjTimestampAdd
+#include "orionld/db/dbModelFromApiSubAttribute.h"               // Own interface
+
+
+
+// -----------------------------------------------------------------------------
+//
+// dbModelFromApiSubAttribute -
+//
+// PARAMETERS
+//   - saP            pointer to the sub-attribute KjNode tree
+//   - dbMdP          array of names od alreadty existing metadata (from DB)
+//   - mdAddedV       array of names of new metadata (sub-attributes)
+//   - mdRemovedV     array of names of sub-attributes that are to be removed (RHS == null)
+//
+bool dbModelFromApiSubAttribute(KjNode* saP, KjNode* dbMdP, KjNode* mdAddedV, KjNode* mdRemovedV)
+{
+  LM_TMP(("KZ3: In dbModelFromApiSubAttribute for sub-attr '%s'", saP->name));
+
+  char* saDotName = kaStrdup(&orionldState.kalloc, saP->name);  // Needed for mdAddedV, mdRemovedV
+
+  dotForEq(saP->name);  // Orion DB-Model states that dots are replaced for equal signs in sub-attribute names
+
+  KjNode* dbSubAttributeP = (dbMdP == NULL)? NULL : kjLookup(dbMdP, saP->name);
+
+  if (saP->type == KjNull)
+  {
+    if (dbSubAttributeP == NULL)
+    {
+      LM_W(("Attempt to DELETE a sub-attribute that doesn't exist (%s)", saP->name));
+      orionldError(OrionldResourceNotFound, "Cannot delete a sub-attribute that does not exist", saP->name, 404);
+      // Add to 207
+      return false;
+    }
+
+    LM_TMP(("KZ: '%s' has a  null RHS - to be removed", saP->name));
+    kjChildAdd(mdRemovedV, kjString(orionldState.kjsonP, NULL, saDotName));
+    return true;
+  }
+
+  if (strcmp(saP->name, "value") == 0)
+  {
+    return true;
+  }
+  else if ((strcmp(saP->name, "object") == 0) || (strcmp(saP->name, "languageMap") == 0))
+  {
+    LM_TMP(("KZ3: Changing 'object' to 'value'"));
+    saP->name = (char*) "value";  // Orion-LD's database model states that all attributes have a "value"
+  }
+  else if (strcmp(saP->name, "observedAt") == 0)
+  {
+    //
+    // observedAt is stored as a JSON object, as any other sub-attribute (my mistake, bad idea but too late now)
+    // Also, the string representation of the ISO8601 is turned into a floating point representation
+    // So, those two problems are fixed here:
+    //
+    double timestamp = parse8601Time(saP->value.s);
+    KjNode* valueP   = kjFloat(orionldState.kjsonP, "value", timestamp);
+
+    saP->type = KjObject;
+    saP->value.firstChildP = valueP;
+    saP->lastChild         = valueP;
+
+    // Also, these two special sub-attrs have no creDate/modDate but they're still sub-attrs, need to be id "mdNames"
+    if (dbSubAttributeP == NULL)
+      kjChildAdd(mdAddedV, kjString(orionldState.kjsonP, NULL, saP->name));
+  }
+  else if (strcmp(saP->name, "unitCode") == 0)
+  {
+    //
+    // unitCode is stored as a JSON object, as any other sub-attribute (my mistake, bad idea but too late now)
+    // So, that problem is fixed here:
+    //
+    KjNode* valueP = kjString(orionldState.kjsonP, "value", saP->value.s);
+
+    saP->type = KjObject;
+    saP->value.firstChildP = valueP;
+    saP->lastChild         = valueP;
+
+    // Also, these two special sub-attrs have no creDate/modDate but they're still sub-attrs, need to be id "mdNames"
+    if (dbSubAttributeP == NULL)
+      kjChildAdd(mdAddedV, kjString(orionldState.kjsonP, NULL, saP->name));
+  }
+  else
+  {
+    LM_TMP(("KZ4: THIS is a sub-attr (%s) - change to 'value' if 'type == object/languageMap", saP->name));
+
+    // If object or languageMap exist, change name to value
+    KjNode* valueP = kjLookup(saP, "object");
+
+    if (valueP == NULL)
+      valueP = kjLookup(saP, "languageMap");
+
+    if (valueP != NULL)
+      valueP->name = (char*) "value";
+
+    // If the sub-attr didn't exist, it needs a "createdAt"
+    if (dbSubAttributeP == NULL)
+    {
+      kjTimestampAdd(saP, "createdAt");
+
+      // The "mdNames" array stores the sub-attribute names in their original names, with dots - saDotName
+      kjChildAdd(mdAddedV, kjString(orionldState.kjsonP, NULL, saDotName));
+    }
+
+    // All sub-attrs get a "modifiedAt"
+    kjTimestampAdd(saP, "modifiedAt");
+  }
+
+  return true;
+}
