+++ src/lib/orionld/serviceRoutines/orionldPatchEntity2.cpp
@@ -24,12 +24,10 @@
 */
 extern "C"
 {
-#include "kbase/kMacros.h"                                       // K_VEC_SIZE
 #include "kjson/KjNode.h"                                        // KjNode
 #include "kjson/kjLookup.h"                                      // kjLookup
 #include "kjson/kjBuilder.h"                                     // kjChildRemove
 #include "kjson/kjRender.h"                                      // kjFastRender
-#include "kalloc/kaStrdup.h"                                     // kaStrdup
 }
 
 #include "logMsg/logMsg.h"                                       // LM_*
@@ -42,9 +40,15 @@ extern "C"
 #include "orionld/common/dotForEq.h"                             // dotForEq
 #include "orionld/common/eqForDot.h"                             // eqForDot
 #include "orionld/types/OrionldHeader.h"                         // orionldHeaderAdd
+#include "orionld/kjTree/kjTreeLog.h"                            // kjTreeLog
+#include "orionld/kjTree/kjTimestampAdd.h"                       // kjTimestampAdd
+#include "orionld/kjTree/kjArrayAdd.h"                           // kjArrayAdd
 #include "orionld/kjTree/kjStringValueLookupInArray.h"           // kjStringValueLookupInArray
 #include "orionld/payloadCheck/pCheckEntity.h"                   // pCheckEntity
+#include "orionld/db/dbModelFromApiEntity.h"                     // dbModelFromApiEntity
 #include "orionld/serviceRoutines/orionldPatchEntity.h"          // Own Interface
+
+
 
 //
 // All db functions (all new ones anyway) should return an entity from DB AS IS, i.e. include the database model
@@ -84,390 +88,12 @@ extern "C"
 //      }
 //    ]
 //
-//   The PATCH-function (not DB dependant) pinpoints every node in the tree where there is s difference
+//   The PATCH-function (not DB dependent) pinpoints every node in the tree where there is s difference
 //   The PATH is the pointer to the differing node.
 //   The TREE is the new value for the differing node.
 //
 //   TREE.type == KjNull means removal
 //
-
-
-
-// -----------------------------------------------------------------------------
-//
-// treePresent -
-//
-static void treePresent(KjNode* tree, const char* msg)
-{
-  char buf[2048];  // Make the buffer size a function parameter and use kaAlloc()
-  kjFastRender(tree, buf);
-  LM_TMP(("%s: %s", msg, buf));
-}
-
-
-
-// -----------------------------------------------------------------------------
-//
-// timestampAdd -
-//
-static void timestampAdd(KjNode* containerP, const char* name)
-{
-  KjNode* nodeP = kjFloat(orionldState.kjsonP, name, orionldState.requestTime);
-  kjChildAdd(containerP, nodeP);
-}
-
-
-
-// -----------------------------------------------------------------------------
-//
-// arrayAdd -
-//
-static KjNode* arrayAdd(KjNode* containerP, const char* name)
-{
-  KjNode* nodeP = kjArray(orionldState.kjsonP, name);
-  kjChildAdd(containerP, nodeP);
-  return nodeP;
-}
-
-
-
-// -----------------------------------------------------------------------------
-//
-// dbModelFromApiSubAttribute -
-//
-// PARAMETERS
-//   - saP            pointer to the sub-attribute KjNode tree
-//   - dbMdP          array of names od alreadty existing metadata (from DB)
-//   - mdAddedV       array of names of new metadata (sub-attributes)
-//   - mdRemovedV     array of names of sub-attributes that are to be removed (RHS == null)
-//
-bool dbModelFromApiSubAttribute(KjNode* saP, KjNode* dbMdP, KjNode* mdAddedV, KjNode* mdRemovedV)
-{
-  LM_TMP(("KZ3: In dbModelFromApiSubAttribute for sub-attr '%s'", saP->name));
-
-  char* saDotName = kaStrdup(&orionldState.kalloc, saP->name);  // Needed for mdAddedV, mdRemovedV
-
-  dotForEq(saP->name);  // Orion DB-Model states that dots are replaced for equal signs in sub-attribute names
-
-  KjNode* dbSubAttributeP = (dbMdP == NULL)? NULL : kjLookup(dbMdP, saP->name);
-
-  if (saP->type == KjNull)
-  {
-    if (dbSubAttributeP == NULL)
-    {
-      LM_W(("Attempt to DELETE a sub-attribute that doesn't exist (%s)", saP->name));
-      orionldError(OrionldResourceNotFound, "Cannot delete a sub-attribute that does not exist", saP->name, 404);
-      // Add to 207
-      return false;
-    }
-
-    LM_TMP(("KZ: '%s' has a  null RHS - to be removed", saP->name));
-    kjChildAdd(mdRemovedV, kjString(orionldState.kjsonP, NULL, saDotName));
-    return true;
-  }
-
-  if (strcmp(saP->name, "value") == 0)
-  {
-    return true;
-  }
-  else if ((strcmp(saP->name, "object") == 0) || (strcmp(saP->name, "languageMap") == 0))
-  {
-    LM_TMP(("KZ3: Changing 'object' to 'value'"));
-    saP->name = (char*) "value";  // Orion-LD's database model states that all attributes have a "value"
-  }
-  else if (strcmp(saP->name, "observedAt") == 0)
-  {
-    //
-    // observedAt is stored as a JSON object, as any other sub-attribute (my mistake, bad idea but too late now)
-    // Also, the string representation of the ISO8601 is turned into a floating point representation
-    // So, those two problems are fixed here:
-    //
-    double timestamp = parse8601Time(saP->value.s);
-    KjNode* valueP   = kjFloat(orionldState.kjsonP, "value", timestamp);
-
-    saP->type = KjObject;
-    saP->value.firstChildP = valueP;
-    saP->lastChild         = valueP;
-
-    // Also, these two special sub-attrs have no creDate/modDate but they're still sub-attrs, need to be id "mdNames"
-    if (dbSubAttributeP == NULL)
-      kjChildAdd(mdAddedV, kjString(orionldState.kjsonP, NULL, saP->name));
-  }
-  else if (strcmp(saP->name, "unitCode") == 0)
-  {
-    //
-    // unitCode is stored as a JSON object, as any other sub-attribute (my mistake, bad idea but too late now)
-    // So, that problem is fixed here:
-    //
-    KjNode* valueP = kjString(orionldState.kjsonP, "value", saP->value.s);
-
-    saP->type = KjObject;
-    saP->value.firstChildP = valueP;
-    saP->lastChild         = valueP;
-
-    // Also, these two special sub-attrs have no creDate/modDate but they're still sub-attrs, need to be id "mdNames"
-    if (dbSubAttributeP == NULL)
-      kjChildAdd(mdAddedV, kjString(orionldState.kjsonP, NULL, saP->name));
-  }
-  else
-  {
-    LM_TMP(("KZ4: THIS is a sub-attr (%s) - change to 'value' if 'type == object/languageMap", saP->name));
-
-    // If object or languageMap exist, change name to value
-    KjNode* valueP = kjLookup(saP, "object");
-
-    if (valueP == NULL)
-      valueP = kjLookup(saP, "languageMap");
-
-    if (valueP != NULL)
-      valueP->name = (char*) "value";
-
-    // If the sub-attr didn't exist, it needs a "createdAt"
-    if (dbSubAttributeP == NULL)
-    {
-      timestampAdd(saP, "createdAt");
-
-      // The "mdNames" array stores the sub-attribute names in their original names, with dots - saDotName
-      kjChildAdd(mdAddedV, kjString(orionldState.kjsonP, NULL, saDotName));
-    }
-
-    // All sub-attrs get a "modifiedAt"
-    timestampAdd(saP, "modifiedAt");
-  }
-
-  return true;
-}
-
-
-
-// -----------------------------------------------------------------------------
-//
-// dbModelFromApiAttribute -
-//
-// FOREACH attribute:
-//   * First an object "md" is created, and all fields of the attribute, except the special ones are moved inside "md".
-//     Special fields:
-//     - type
-//     - value/object/languageMap (must be renamed to "value" - that's part of the database model)
-//     - observedAt
-//     - datasetId
-//     - unitCode
-//
-//   * Actually, all children of 'attrP' are moved to 'md'
-//   * And then, all the special fields are moved back to 'attrP'
-//   * ".added"   is added
-//   * ".removed" is added
-//   * "modDate"  is added
-//   * "creDate"  is added iff the attribute did not previously exist
-//
-bool dbModelFromApiAttribute(KjNode* attrP, KjNode* dbAttrsP, KjNode* attrAddedV, KjNode* attrRemovedV)
-{
-  KjNode* mdP         = NULL;
-  char*   attrDotName = kaStrdup(&orionldState.kalloc, attrP->name);  // Needed for attrAddedV, attrRemovedV
-
-  dotForEq(attrP->name);  // Orion DB-Model states that dots are replaced for equal signs in attribute names
-
-  if (attrP->type == KjNull)
-  {
-    LM_TMP(("KZ: '%s' has a null RHS - to be removed", attrP->name));
-    KjNode* attrNameP = kjString(orionldState.kjsonP, NULL, attrDotName);
-    kjChildAdd(attrRemovedV, attrNameP);
-    return true;
-  }
-
-  // Move everything into "md", leaving attrP EMPTY
-  mdP = kjObject(orionldState.kjsonP, "md");
-  mdP->value.firstChildP   = attrP->value.firstChildP;
-  mdP->lastChild           = attrP->lastChild;
-  attrP->value.firstChildP = NULL;
-  attrP->lastChild         = NULL;
-
-  // Move special fields back to "attrP"
-  const char* specialV[] = { "type", "value", "object", "languageMap", "datasetId" };  // observedAt+unitCode are mds (db-model)
-  for (unsigned int ix = 0; ix < K_VEC_SIZE(specialV); ix++)
-  {
-    KjNode* nodeP = kjLookup(mdP, specialV[ix]);
-    if (nodeP != NULL)
-    {
-      if ((ix == 2) || (ix == 3))         // "object", "languageMap", change name to "value" (Orion's DB model)
-        nodeP->name = (char*) "value";
-
-      kjChildRemove(mdP, nodeP);
-      kjChildAdd(attrP, nodeP);
-    }
-  }
-
-  // Also, the attribute is being modified - need to update the "modifiedAt" (called modDate in Orion's DB-Model)
-  timestampAdd(attrP, "modDate");
-
-  KjNode*  mdAddedP     = NULL;
-  KjNode*  mdRemovedP   = NULL;
-  KjNode*  mdNamesP     = NULL;
-  KjNode*  dbAttrP      = kjLookup(dbAttrsP, attrP->name);
-  KjNode*  dbMdP        = NULL;
-
-  if (dbAttrP == NULL)  // Attribute does not already exist
-  {
-    if (attrP->type == KjNull)
-    {
-      LM_W(("Attempt to DELETE an attribute that doesn't exist (%s)", attrP->name));
-      orionldError(OrionldResourceNotFound, "Cannot delete an attribute that does not exist", attrP->name, 404);
-      // Add to 207
-      return false;
-    }
-
-    timestampAdd(attrP, "creDate");
-
-    // The "attrNames" array stores the sub-attribute names in their original names, with dots - attrDotName
-    kjChildAdd(attrAddedV, kjString(orionldState.kjsonP, NULL, attrDotName));
-
-    // All attributes have a field 'mdNames' - it's an array
-    mdNamesP = kjArray(orionldState.kjsonP, "mdNames");
-    kjChildAdd(attrP, mdNamesP);
-
-    LM_TMP(("KX: the attribute is NEW - if it has any sub-attributes, the entire 'md' needs to be included, never mind .added, .removed etc"));
-  }
-  else
-  {
-    // The attribute is not new, but, perhaps it didn't have any sub-attrs before but now it does?
-    // If so, never mind .added, .removed etc, what we need is to include the entire "md"
-    //
-    dbMdP    = kjLookup(dbAttrP, "md");
-    mdNamesP = kjLookup(dbAttrP, "mdNames");
-
-    if (mdNamesP == NULL)
-    {
-      //
-      // No 'mdNames' field in DB Attribute?
-      // That's a DB Error.
-      // Never mind, let's just silently fix it
-      //
-      mdNamesP = kjArray(orionldState.kjsonP, ".names");
-    }
-    else
-    {
-      kjChildRemove(dbAttrP, mdNamesP);
-      mdNamesP->name = (char*) ".names";  // We hijack the 'mdNames' array for the compilation of the Patch Tree
-    }
-
-    kjChildAdd(attrP, mdNamesP);
-
-    mdAddedP   = arrayAdd(attrP, ".added");    // FIXME: Should not be needed if dbMdP == NULL
-    mdRemovedP = arrayAdd(attrP, ".removed");  // FIXME: Should not be needed if dbMdP == NULL
-  }
-
-  treePresent(attrP, "KZ4: Attribute with unprocessed sub-attributes");
-  if (attrP->type == KjNull)
-    return true;
-
-  // Sub-Attributes
-  LM_TMP(("KZ: Loop over sub-attributes (contents of 'mdP')"));
-  for (KjNode* subP = mdP->value.firstChildP; subP != NULL; subP = subP->next)
-  {
-    LM_TMP(("KZ5: sub-attribute '%s'", subP->name));
-    dbModelFromApiSubAttribute(subP, dbMdP, mdAddedP, mdRemovedP);
-  }
-
-  if (mdP->value.firstChildP != NULL)
-  {
-    LM_TMP(("KY: putting 'md' inside attr"));
-    kjChildAdd(attrP, mdP);
-  }
-  treePresent(attrP, "KY: Attribute with PROCESSED sub-attributes");
-
-  return true;
-}
-
-
-
-// -----------------------------------------------------------------------------
-//
-// dbModelFromApiEntity - modify the request tree to match the db model
-//
-// Modifications:
-//
-//   * Entity Level (dbModelEntity)
-//     * "id" can't be modified    (make sure it's removed and partial error reported by pCheckEntity)
-//     * "type" can't be modified  (make sure it's removed and partial error reported by pCheckEntity)
-//     * "scope" doesn't exist yet (make sure it's removed and partial error reported by pCheckEntity)
-//     * "modDate" is set with orionldState.requestTime
-//     * "attrs" member is created and added
-//     * It's an attribute - move to "attrs" and call "attributeToDbModel" on it.
-//     * if Array - datasetId - error for now (partial 501)
-//     * if != Object - error (as pCheckAttribute makes sure the request tree is normalized)
-//     * move the attribute there + call "Level 1 Function" (orionldDbModelAttribute)
-//
-//   * Attribute Level (dbModelAttribute)
-//     * If Array, recursive call for each member (set the name to the sttribute name)
-//     * "datasetId" present - call orionldDbModelAttributeDatasetId
-//     * "type" can't be modified
-//     * "value"/"object"/"languageMap" changes name to "value" and RHS stays as is
-//     * "observedAt" is made an Object with single member "value"
-//     * "unitCode" is made an Object with single member "value"
-//     * "md" is created and added
-//
-//   * Sub-Attribute Level (dbModelSubAttribute)
-//
-bool dbModelFromApiEntity(KjNode* entityP, KjNode* dbAttrsP, KjNode* dbAttrNamesP)
-{
-  KjNode*      nodeP;
-  const char*  mustGo[] = { "_id", "id", "@id", "type", "@type", "scope", "createdAt", "modifiedAt", "creDate", "modDate" };
-
-  //
-  // Remove any non-attribute nodes
-  // PATCH-able toplevel fields (type? scope?) would need to be removed and saved. Later reintroduced into entityP
-  //
-  for (unsigned int ix = 0; ix < sizeof(mustGo) / sizeof(mustGo[0]); ix++)
-  {
-    nodeP = kjLookup(entityP, mustGo[ix]);
-    if (nodeP != NULL)
-      kjChildRemove(entityP, nodeP);
-  }
-
-
-  //
-  // Create the "attrs" field and move all attributes from entityP to there ("attrs")
-  // Then move "attrs" inside entityP
-  //
-  KjNode* attrsP = kjObject(orionldState.kjsonP, "attrs");
-  attrsP->value.firstChildP = entityP->value.firstChildP;
-  attrsP->lastChild         = entityP->lastChild;
-
-  // Make "attrs" the only child (thus far) of entityP
-  entityP->value.firstChildP  = attrsP;
-  entityP->lastChild          = attrsP;
-
-  // Rename "attrNames" to ".names" and add it to the entity
-  if (dbAttrNamesP == NULL)
-  {
-    // No 'attrNames' field in DB Entity - must be the entity had zero attrs before this
-    dbAttrNamesP = kjArray(orionldState.kjsonP, ".names");
-  }
-  else
-    dbAttrNamesP->name = (char*) ".names";
-
-  // Now we can add "attrNames"
-  kjChildAdd(entityP, dbAttrNamesP);
-
-  // Adding members necessary for the DB Model
-  timestampAdd(entityP, "modDate");
-  KjNode* attrAddedV   = arrayAdd(entityP, ".added");
-  KjNode* attrRemovedV = arrayAdd(entityP, ".removed");  // Not really necessary ... the RHS == null already tells the next layer
-
-  treePresent(entityP, "KZ: DB-prepped Entity with unprocessed attributes");
-
-  //
-  // Loop over the "attrs" member of entityP and call dbModelFromApiAttribute for every attribute
-  //
-  for (KjNode* attrP = attrsP->value.firstChildP; attrP != NULL; attrP = attrP->next)
-  {
-    dbModelFromApiAttribute(attrP, dbAttrsP, attrAddedV, attrRemovedV);
-  }
-
-  treePresent(entityP, "KY: DB-prepped Entity with PROCESSED attributes");
-
-  return true;
-}
 
 
 
@@ -496,19 +122,10 @@ void patchTreeItemAdd(KjNode* patchTree, const char* path, KjNode* valueP, const
 
 
 
-static void namesArrayAddedToPatchTree(KjNode* patchTree, const char* path, KjNode* addedP)
-{
-  LM_TMP(("KZ2: Something added, nothing removed - $push to array"));
-  patchTreeItemAdd(patchTree, path, addedP, "PUSH");
-}
-
-static void namesArrayRemovedToPatchTree(KjNode* patchTree, const char* path, KjNode* removedP)
-{
-  LM_TMP(("KZ2: Nothing added, somehing removed - $pull from array"));
-
-  patchTreeItemAdd(patchTree, path, removedP, "PULL");
-}
-
+// -----------------------------------------------------------------------------
+//
+// namesArrayMergeToPatchTree -
+//
 static void namesArrayMergeToPatchTree(KjNode* patchTree, const char* path, KjNode* namesP, KjNode* addedP, KjNode* removedP)
 {
   LM_TMP(("KZ2: Something added, something removed - merge and $set array"));
@@ -533,13 +150,15 @@ static void namesArrayMergeToPatchTree(KjNode* patchTree, const char* path, KjNo
 
 
 
+// -----------------------------------------------------------------------------
+//
+// namesToPatchTree -
+//
 static void namesToPatchTree(KjNode* patchTree, const char* path, KjNode* namesP, KjNode* addedP, KjNode* removedP)
 {
-  LM_TMP(("KZ2 -------------- Found a names array. path: '%s'", path));
-
-  treePresent(namesP,   "KZ2: names");
-  treePresent(addedP,   "KZ2: added");
-  treePresent(removedP, "KZ2: removed");
+  kjTreeLog(namesP,   "KZ2: names");
+  kjTreeLog(addedP,   "KZ2: added");
+  kjTreeLog(removedP, "KZ2: removed");
 
   // Fix the attrNames/mdNames $set/$pull/$pop
   char* namesPath;
@@ -554,18 +173,15 @@ static void namesToPatchTree(KjNode* patchTree, const char* path, KjNode* namesP
     snprintf(namesPath, namesPathLen, "%s.mdNames", path);
   }
 
-  LM_TMP(("KZ2: Fix %s", namesPath));
-
   bool added   = false;
   bool removed = false;
 
   if ((addedP   != NULL) && (addedP->value.firstChildP   != NULL))  added   = true;
   if ((removedP != NULL) && (removedP->value.firstChildP != NULL))  removed = true;
 
-  if      ((added == false)  && (removed == false))   LM_TMP(("KZ2: Nothing added, nothing removed - nothing needs to be done!"));
-  else if ((added == true)   && (removed == false))   namesArrayAddedToPatchTree(patchTree,   namesPath, addedP);
-  else if ((added == false)  && (removed == true))    namesArrayRemovedToPatchTree(patchTree, namesPath, removedP);
-  else if ((added == true)   && (removed == true))    namesArrayMergeToPatchTree(patchTree,   namesPath, namesP, addedP, removedP);
+  if      ((added == true)   && (removed == false))   patchTreeItemAdd(patchTree, namesPath, addedP,   "PUSH");
+  else if ((added == false)  && (removed == true))    patchTreeItemAdd(patchTree, namesPath, removedP, "PULL");
+  else if ((added == true)   && (removed == true))    namesArrayMergeToPatchTree(patchTree, namesPath, namesP, addedP, removedP);
 }
 
 
@@ -587,33 +203,25 @@ static void namesToPatchTree(KjNode* patchTree, const char* path, KjNode* namesP
 //
 void orionldEntityPatchTree(KjNode* oldP, KjNode* newP, char* path, KjNode* patchTree)
 {
-  // <DEBUG>
   LM_TMP(("KZ: path: %s", path));
-  LM_TMP(("KZ: old:  %s", (oldP == NULL)? "NULL" : oldP->name));
-  LM_TMP(("KZ: new:  %s", (newP == NULL)? "NULL" : newP->name));
-  // </DEBUG>
 
   if (newP == NULL)  // Not sure this is ever a possibility ...
     return;
 
-  if (oldP == NULL)
+  if (oldP == NULL)  // It's NEW - didn't exist - simply ADD
   {
-    char buf[1024];
-    kjFastRender(newP, buf);
-    LM_TMP(("KZ: patchTreeItemAdd(PATH: %s, TREE: %s)", path, buf));
     patchTreeItemAdd(patchTree, path, newP, NULL);
     return;
   }
 
-  // newP != NULL && oldP != NULL
-  if (newP->type == KjNull)
+  if (newP->type == KjNull)  // NULL as RHS means REMOVAL
   {
-    LM_TMP(("KZ: patchTreeItemAdd(PATH: %s, TREE: null - REMOVAL)", path));
+    LM_TMP(("KZ: patchTreeItemAdd(PATH: %s, TREE: null - FOR REMOVAL)", path));
     patchTreeItemAdd(patchTree, path, newP, NULL);
     return;
   }
 
-  if (newP->type != oldP->type)
+  if (newP->type != oldP->type)  // Different type?  - overwrite the old
   {
     LM_TMP(("KZ: patchTreeItemAdd(PATH: %s, TREE: <JSON %s>)", path, kjValueType(newP->type)));
     patchTreeItemAdd(patchTree, path, newP, NULL);
@@ -621,7 +229,8 @@ void orionldEntityPatchTree(KjNode* oldP, KjNode* newP, char* path, KjNode* patc
   }
 
   bool change = false;
-  bool leave  = false;
+  bool leave  = false;  // We're done if it's a SIMPLE TYPE (String, Number, Bool)
+
   if      (newP->type == KjString)   { if (strcmp(newP->value.s, oldP->value.s) != 0)  change = true; leave = true; }
   else if (newP->type == KjInt)      { if (newP->value.i != oldP->value.i)             change = true; leave = true; }
   else if (newP->type == KjFloat)    { if (newP->value.f != oldP->value.f)             change = true; leave = true; }
@@ -636,17 +245,18 @@ void orionldEntityPatchTree(KjNode* oldP, KjNode* newP, char* path, KjNode* patc
   if (leave == true)
     return;
 
-  if (newP->type == KjArray)
+  if (newP->type == KjArray)  // If it's an Array, we replace the old array (by definition)
   {
     LM_TMP(("KZ: patchTreeItemAdd(PATH: %s, TREE: <JSON %s>)", path, kjValueType(newP->type)));
     patchTreeItemAdd(patchTree, path, newP, NULL);
     return;
   }
 
-  // Both are JSON Object - entering
-  // Careful about the newP linked lists - the recursive calls may call patchTreeItemAdd and that removes items from the list
+  // Both are JSON Object - entering to find differences
+
   //
-  // Can't use the normal "for" - must be a "while and a next pointer"
+  // Careful with the linked lists of newP - the recursive calls may invoke patchTreeItemAdd and that REMOVES ITEMS FROM A LIST
+  // => Can't use the normal "for" - must be a "while and a next pointer"
   //
   KjNode* newItemP   = newP->value.firstChildP;
   KjNode* next;
@@ -701,17 +311,6 @@ void orionldEntityPatchTree(KjNode* oldP, KjNode* newP, char* path, KjNode* patc
 //
 // orionldPatchEntity2 -
 //
-// For all this Real PATCH thingy to work for any API operation, we will need the following
-//
-//   * mongocEntitiesGet() - returning entities as they are in the DB - with "attrs", "attrNames", and "$datasets" children as objects of the output array
-//
-//   * orionldEntityPatch(KjNode* dbEntity, KjNode* patchEntity)
-//   * patchAttribute(KjNode* dbAttribute, KjNode* patchAttribute, KjNode* dbAttrNamesArray, bson_t* toSet, bson_t* toUnset)
-//   * patchSubAttribute(KjNode* dbSubAttribute, KjNode* patchSubAttribute, KjNode* dbMdNamesArray, char* dbPath, bson_t* toSet, bson_t* toUnset)
-//
-//
-//
-//
 bool orionldPatchEntity2(void)
 {
   if (experimental == false)
@@ -722,7 +321,6 @@ bool orionldPatchEntity2(void)
 
   char*    entityId    = orionldState.wildcard[0];
   KjNode*  dbEntityP;
-  char     buf[4096];  // DEBUG
 
   //
   // FIXME: would really like to only extract the attrs in the patch ...
@@ -755,11 +353,7 @@ bool orionldPatchEntity2(void)
     return false;
   }
 
-  kjFastRender(dbEntityP, buf);
-  LM_TMP(("DB Entity:  %s", buf));
-  kjFastRender(orionldState.requestTree, buf);
-  LM_TMP(("Patch Tree: %s", buf));
-
+  //
   // FIXME: If I destroy the tree, how do I handle forwarding later ... ?
   //        I extract the part that is to be forwarded from the tree before I destroy
   //        If non-exclusive, might be I both forward and do it locally - kjClone(attrP)
@@ -768,10 +362,9 @@ bool orionldPatchEntity2(void)
   dbModelFromApiEntity(orionldState.requestTree, dbAttrsP, dbAttrNamesP);
 
   //
-  // orionldEntityPatch - patching the entity
-  //
-  // Due to the questionable database model of Orion, that Orion-LD has inherited (backwards compatibility),
+  // Due to the questionable database model of Orion (that Orion-LD has inherited - backwards compatibility),
   // the attributes that have been added/removed must be added/removed to a redundant database field called "attrNames".
+  //
   // It's not possible to both remove and add items from a field in a single operation in mongo, so a list of all
   // added attributes and another list of all deleted attributes must be maintained for the purpose of updating "attrNames".
   //
@@ -785,16 +378,10 @@ bool orionldPatchEntity2(void)
   //
   KjNode* patchTree     = kjArray(orionldState.kjsonP, NULL);
   KjNode* dbAttrsObject = kjObject(orionldState.kjsonP, NULL);
-  kjChildAdd(dbAttrsObject, dbAttrsP);
 
-  LM_TMP(("KY: --------------------------- Before orionldEntityPatchTree --------------------------"));
-  treePresent(dbAttrsObject, "KY: DB-Attrs");
-  treePresent(orionldState.requestTree, "KY: Patch");
-  LM_TMP(("KZ: --------------------------- Calling orionldEntityPatchTree -------------------------"));
+  kjChildAdd(dbAttrsObject, dbAttrsP);
   orionldState.requestTree->name = NULL;
   orionldEntityPatchTree(dbAttrsObject, orionldState.requestTree, NULL, patchTree);
-  LM_TMP(("KZ2: --------------------------- After orionldEntityPatchTree ---------------------------"));
-  treePresent(patchTree, "KZ2: PATCH TREE");
 
   bool b = mongocEntityUpdate(entityId, patchTree);  // Added/Removed (sub-)attrs are found as arrays named ".added" and ".removed"
   if (b == false)
@@ -803,10 +390,9 @@ bool orionldPatchEntity2(void)
 
     LM_E(("mongocEntityUpdate(%s): [%d.%d]: %s", entityId, errP->domain, errP->code, errP->message));
 
-    if (errP->code == 12)
-      orionldError(OrionldResourceNotFound, "Entity not found", entityId, 404);
-    else
-      orionldError(OrionldInternalError, "Internal Error", errP->message, 500);
+    if (errP->code == 12)  orionldError(OrionldResourceNotFound, "Entity not found", entityId, 404);
+    else                   orionldError(OrionldInternalError, "Internal Error", errP->message, 500);
+
     return false;
   }
