+++ test/functionalTest/cases/0000_ngsild/ngsild_entity_with_geoproperty_with_subattrs.test
@@ -1,0 +1,145 @@
+# Copyright 2022 FIWARE Foundation e.V.
+#
+# This file is part of Orion-LD Context Broker.
+#
+# Orion-LD Context Broker is free software: you can redistribute it and/or
+# modify it under the terms of the GNU Affero General Public License as
+# published by the Free Software Foundation, either version 3 of the
+# License, or (at your option) any later version.
+#
+# Orion-LD Context Broker is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero
+# General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with Orion-LD Context Broker. If not, see http://www.gnu.org/licenses/.
+#
+# For those usages not covered by this license please contact with
+# orionld at fiware dot org
+
+# VALGRIND_READY - to mark the test ready for valgrindTestSuite.sh
+
+--NAME--
+Entity with GeoProperty having sub-attributes - checking their compaction in GET after a PATCH
+
+--SHELL-INIT--
+export BROKER=orionld
+dbInit CB
+brokerStart CB 212-249
+
+--SHELL--
+
+#
+# 01. Create an entity E1 with a Property 'temperature' that has a sub-Relationship 'providedBy'
+# 02. Retrieve the entity
+# 03. PATCH the attributes of E1
+# 04. Retrieve the entity
+#
+
+echo "01. Create an entity E1 with a Property 'temperature' that has a sub-Relationship 'providedBy'"
+echo "=============================================================================================="
+payload='{
+  "id": "urn:E1",
+  "type": "T",
+  "temperature": {
+    "value": 7,
+    "providedBy": {
+      "object": "urn:ngsi-ld:Building:barn001"
+    }
+  }
+}'
+orionCurl --url /ngsi-ld/v1/entities -X POST --payload "$payload"
+echo
+echo
+
+
+echo "02. Retrieve the entity"
+echo "======================="
+orionCurl --url /ngsi-ld/v1/entities/urn:E1?options=concise
+echo
+echo
+
+
+echo "03. PATCH the attributes of E1"
+echo "=============================="
+payload='{
+  "temperature": {
+    "value": 7,
+    "providedBy": {
+      "object": "urn:ngsi-ld:Building:barn001"
+    }
+  }
+}'
+orionCurl --url /ngsi-ld/v1/entities/urn:E1/attrs -X PATCH --payload "$payload"
+echo
+echo
+
+
+echo "04. Retrieve the entity"
+echo "======================="
+orionCurl --url /ngsi-ld/v1/entities/urn:E1?options=concise
+echo
+echo
+
+
+--REGEXPECT--
+01. Create an entity E1 with a Property 'temperature' that has a sub-Relationship 'providedBy'
+==============================================================================================
+HTTP/1.1 201 Created
+Content-Length: 0
+Location: /ngsi-ld/v1/entities/urn:E1
+Date: REGEX(.*)
+
+
+
+02. Retrieve the entity
+=======================
+HTTP/1.1 200 OK
+Content-Length: 107
+Content-Type: application/json
+Link: <https://uri.etsi.org/ngsi-ld/v1/ngsi-ld-core-context.jsonld>; rel="http://www.w3.org/ns/json-ld#context"; type="application/ld+json"
+Date: REGEX(.*)
+
+{
+    "id": "urn:E1",
+    "temperature": {
+        "providedBy": {
+            "object": "urn:ngsi-ld:Building:barn001"
+        },
+        "value": 7
+    },
+    "type": "T"
+}
+
+
+03. PATCH the attributes of E1
+==============================
+HTTP/1.1 204 No Content
+Date: REGEX(.*)
+
+
+
+04. Retrieve the entity
+=======================
+HTTP/1.1 200 OK
+Content-Length: 107
+Content-Type: application/json
+Link: <https://uri.etsi.org/ngsi-ld/v1/ngsi-ld-core-context.jsonld>; rel="http://www.w3.org/ns/json-ld#context"; type="application/ld+json"
+Date: REGEX(.*)
+
+{
+    "id": "urn:E1",
+    "temperature": {
+        "providedBy": {
+            "object": "urn:ngsi-ld:Building:barn001"
+        },
+        "value": 7
+    },
+    "type": "T"
+}
+
+
+--TEARDOWN--
+brokerStop CB
+dbDrop CB
